+++ src-tauri/src/lib.rs
@@ -53,6 +53,7 @@ pub fn run() {
             commands::thumbnail::get_file_sizes,
             commands::thumbnail::get_image_dimensions,
             commands::video::compress_video,
+            commands::video::trim_video,
             commands::video::preview_video_compression,
             commands::video::cancel_video_compress,
             commands::video::check_ffmpeg,